requirements.txt
Django==1.11.9
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
psycopg2==2.7.3.2
pytz==2017.3
